requirements.txt
pytest==8.3.1
allure-pytest==2.13.5
selene>=2.0.0rc9
python-dotenv==1.0.1
pydantic==2.7.3
pydantic-settings==2.3.4
black==24.8.0